Moved all the incident related logic to separate files from __init__.py Also changed the imports a bit
