+++ app/__init__.py
@@ -1,7 +1,8 @@
 from app.alerts import alert_handle, alert_handle_create, alert_handle_update
 from app.im.mattermost.buttons import mattermost_buttons_handler
 from app.im.slack.buttons import slack_buttons_handler
-from app.incident import Incident, Incidents, create_or_load_incidents
+from app.incident.helpers import create_or_load_incidents
+from app import Incident, Incidents
 from app.logging import logger
 from app.queue import Queue, queue_handle, recreate_queue
 from app.route import generate_route
